btshim: don't restart BLE stack on app exit when Bluetooth is disabled
After the heap guard makes BLE Spam/Walk fail cleanly, the scene's exit handler
calls ble_spam_hal_stop() -> bt_start_stack(), which unconditionally brought the
btshim serial stack back up. But BT is off by default on this no-PSRAM board and
the stack was never running, so "restoring" it (a) is wrong and (b) OOMs: the
controller + Bluedroid need ~64 KB, only ~46 KB is free (measured), so init
aborts inside ESP-IDF -- either BTU_StartUp -> vQueueDelete(NULL) or the
controller link-layer reset tripping the interrupt watchdog.

bt_handle_start_stack now returns early when bt_settings.enabled is false, so
bt_start_stack is a no-op unless the user actually turned BT on. bt_set_settings
still starts the stack on the off->on toggle.

Also fixed the %zu log format (nano-printf prints it literally and drops the
arg) -> %u with cast.

Verified on hardware: tapping BLE Spam repeatedly now returns to the menu each
time with "Not enough RAM for BLE spam: 46340 free, need >= 73728" and no reboot.

+++ applications/main/ble_spam/ble_spam_hal.c
@@ -78,8 +78,8 @@ bool ble_spam_hal_start(void) {
     if(free_internal < BLE_SPAM_MIN_FREE_INTERNAL) {
         ESP_LOGE(
             TAG,
-            "Not enough RAM for BLE spam: %zu free, need >= %u",
-            free_internal,
+            "Not enough RAM for BLE spam: %u free, need >= %u",
+            (unsigned)free_internal,
             (unsigned)BLE_SPAM_MIN_FREE_INTERNAL);
         return false;
     }
